Add additional JSDoc
