+++ packages/user-operation-controller/src/types.ts
@@ -1,3 +1,6 @@
+/**
+ * A complete user operation to be submitted to a bundler.
+ */
 export type UserOperation = {
   /** The data to pass to the sender during the main execution call. */
   callData: string;
@@ -32,7 +35,7 @@ export type UserOperation = {
    */
   paymasterAndData: string;
 
-  /** The amount of gas to pay to compensate the bundler for pre-verification execution, calldata and any gas overhead that can’t be tracked on-chain. */
+  /** The amount of gas to pay to compensate the bundler for pre-verification execution, calldata and any gas overhead that cannot be tracked on-chain. */
   preVerificationGas: string;
 
   /** The account making the operation. */
@@ -45,6 +48,9 @@ export type UserOperation = {
   verificationGasLimit: string;
 };
 
+/**
+ * The possible statuses of a single user operation.
+ */
 export enum UserOperationStatus {
   Unapproved = 'unapproved',
   Approved = 'approved',
@@ -54,6 +60,7 @@ export enum UserOperationStatus {
   Confirmed = 'confirmed',
 }
 
+/** Information concerning an error while creating a user operation. */
 export type UserOperationError = {
   /**
    * A descriptive error name.
@@ -76,87 +83,178 @@ export type UserOperationError = {
   code: string | null;
 
   /**
-   * The rpc property holds additional information related to the error.
+   * Additional information related to the error.
    */
   rpc: string | null;
 };
 
+/**
+ * Metadata concerning a single user operation, stored in the client state.
+ */
 export type UserOperationMetadata = {
+  /** The URL of the bundler that the operation was submitted to. */
   bundlerUrl: string | null;
+
+  /** The hexadecimal chain ID of the target network. */
   chainId: string;
+
+  /** An error that occurred while creating the operation. */
   error: UserOperationError | null;
+
+  /** The hash of the user operation, generated by the bundler. */
   hash: string | null;
+
+  /** A unique ID used to identify a user operation in the client. */
   id: string;
+
+  /** The current status of the user operation. */
   status: UserOperationStatus;
+
+  /** The timestamp of when the user operation was created. */
   time: number;
+
+  /** The resulting user operation object to be submitted to the bundler. */
   userOperation: UserOperation;
 };
 
-export type BundlerEstimateUserOperationGasResponse = {
-  preVerificationGas: number;
-  verificationGas: number;
-  verificationGasLimit: number;
-  callGasLimit: number;
-};
-
-export type UserOperationReceipt = {
-  actualGasCost: string;
-  actualGasUsed: string;
-  success: boolean;
-  receipt: {
-    blockHash: string;
-    transactionHash: string;
-  };
-};
-
+/**
+ * The data provided to the smart contract account when preparing a user operation.
+ */
 export type PrepareUserOperationRequest = {
+  /** The hexadecimal chain ID of the target network. */
   chainId: string;
+
+  /** The data to include in the resulting transaction.  */
   data?: string;
+
+  /** The destination address of the resulting transaction. */
   to?: string;
+
+  /** The value to send in the resulting transaction. */
   value?: string;
 };
 
+/**
+ * The data provided to the smart contract account when updating a user operation.
+ */
 export type UpdateUserOperationRequest = {
+  /** The user operation to update including the dummy signature and dummy paymasterAndData values. */
   userOperation: UserOperation;
 };
 
+/**
+ * The data provided to the smart contract account when signing a user operation.
+ */
 export type SignUserOperationRequest = {
+  /** The user operation to sign including the dummy signature and final paymasterAndData values. */
   userOperation: UserOperation;
+
+  /** The hexadecimal chain ID of the target network. */
   chainId: string;
 };
 
+/**
+ * The data returned by the smart contract account when preparing a user operation.
+ */
 export type PrepareUserOperationResponse = {
+  /** The URL of the bundler to submit the user operation to. */
   bundler: string;
+
+  /** The data to pass to the sender during the main execution call. */
   callData: string;
+
+  /**
+   * A dummy paymasterAndData value required to estimate gas using the bundler.
+   * Only required if the smart contract account is not providing gas values.
+   */
   dummyPaymasterAndData?: string;
+
+  /**
+   * A dummy signature value required to estimate gas using the bundler.
+   * Only required if the smart contract account is not providing gas values.
+   */
   dummySignature?: string;
+
+  /**
+   * The estimated gas limits for the user operation.
+   * Gas is automatically estimated using the bundler if not provided.
+   */
   gas?: {
+    /** The amount of gas to allocate to the main execution call. */
     callGasLimit: string;
+
+    /** The amount of gas to pay to compensate the bundler for pre-verification execution, calldata and any gas overhead that cannot be tracked on-chain. */
     preVerificationGas: string;
+
+    /** The amount of gas to allocate for the verification step. */
     verificationGasLimit: string;
   };
+
+  /**
+   * The initCode to include in the user operation.
+   * Required only if the smart contract is not yet deployed.
+   */
   initCode?: string;
+
+  /**
+   * The nonce to include in the user operation, specific to the smart contract account.
+   */
   nonce: string;
+
+  /**
+   * The address of the smart contract account creating the user operation.
+   */
   sender: string;
 };
 
+/**
+ * The data returned by the smart contract account when updating a user operation.
+ */
 export type UpdateUserOperationResponse = {
+  /**
+   * The final paymasterAndData to include in the user operation.
+   * Not required if a paymaster is not sponsoring the transaction.
+   */
   paymasterAndData?: string;
 };
 
+/**
+ * The data returned by the smart contract account when signing a user operation.
+ */
 export type SignUserOperationResponse = {
+  /** The final signature of the user operation. */
   signature: string;
 };
 
+/**
+ * An abstraction to provide smart contract and paymaster specific data when creating a user operation.
+ * This will typically communicate with an account snap installed in the client.
+ */
 export type SmartContractAccount = {
+  /**
+   * Retrieve the initial values required to create a user operation.
+   * @param request - The data needed by the smart contract account to provide the initial user operation values.
+   * @returns The initial values required to create a user operation.
+   */
   prepareUserOperation: (
     request: PrepareUserOperationRequest,
   ) => Promise<PrepareUserOperationResponse>;
 
+  /**
+   * Retrieve additional data required to create a user operation, such as the paymasterAndData value.
+   * If gas values were not provided in the prepare response, this will be called after estimating gas using the bundler.
+   * @param request - The data needed by the smart contract account to provide the additional user operation values.
+   * @returns The additional values required to create a user operation.
+   */
   updateUserOperation: (
     request: UpdateUserOperationRequest,
   ) => Promise<UpdateUserOperationResponse>;
 
+  /**
+   * Sign the final user operation.
+   * @param request - The data needed by the smart contract account to generate the signature.
+   * @returns The final values required to sign a user operation.
+   */
   signUserOperation: (
     request: SignUserOperationRequest,
   ) => Promise<SignUserOperationResponse>;
